LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/script-src-report-only-policy-works-with-hash-policy-expected.txt


PASS Test that script executes if allowed by proper hash values
FAIL Test that the securitypolicyviolation event is fired assert_equals: expected "script-src-elem" but got "script-src 'nonce-abc'"
